+++ ProCSharpCode.DotNet5/ProCSharpCode/CsharpCollections/GenericCollections.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 
 using static System.Environment;
 using static ProCSharpCode.Utils.CollectionUtils;
+using System.Collections.Immutable;
 
 namespace ProCSharpCode.CSharpCollections
 {
@@ -44,435 +45,6 @@ namespace ProCSharpCode.CSharpCollections
     //   namespace) 
     #endregion
 
-    // Non-Generic Collections
-    // =======================
-    #region Nongeneric collections (System.Collections namespace)
-
-    // The System.Collections Namespace
-    // ================================
-    // When the.NET platform was first released, programmers frequently used the nongeneric collection
-    // classes found within the System.Collections namespace, which contains a set of classes used to manage
-    // and organize large amounts of in-memory data.
-
-    // Table 9-1 Useful Types of System.Collections
-    //---------------------------------------------
-    // System.Collections Class          Meaning in Life Key         Implemented Interfaces
-    // ==============================================================================================================
-    // ArrayList 
-    // ---------
-    // Represents a dynamically sized collection of objects listed in sequential order
-    // -----------------------------------------------------------------------------------------------------
-    // - Implemented Interfaces:
-    //    IList, ICollection, IEnumerable, and ICloneable.
-    // ==============================================================================================================
-    // BitArray
-    // --------
-    // Manages a compact array of bit values, which are represented as Booleans, where  true indicates that 
-    // the bit is on (1) and false indicates the bit is off(0).
-    // -----------------------------------------------------------------------------------------------------
-    // - Implemented Interfaces:
-    //    ICollection, IEnumerable, and ICloneable
-    //    
-    // ==============================================================================================================
-    // Hashtable
-    // --------- 
-    // Represents a collection of key-value pairs that are organized based on the hash code of the key.
-    // -----------------------------------------------------------------------------------------------------
-    // - Implemented Interfaces:
-    //    IDictionary, ICollection, IEnumerable, and ICloneable
-    //    
-    //    
-    // ==============================================================================================================
-    // Queue 
-    // -------- 
-    // Represents a standard first-in, first-out (FIFO) collection of objects.
-    // -----------------------------------------------------------------------------------------------------
-    // - Implemented Interfaces:
-    //    ICollection, IEnumerable,  and ICloneable
-    //   
-    // ==============================================================================================================
-    // SortedList 
-    // ---------- 
-    // Represents a collection of key-value pairs that are sorted by the keys and are accessible
-    // by key and by index
-    // 
-    // -----------------------------------------------------------------------------------------------------
-    // - Implemented Interfaces:
-    //    IDictionary, ICollection,IEnumerable, and ICloneable
-    //    
-    //    
-    // ==============================================================================================================
-    // Stack 
-    // ------
-    // A last-in, first-out (LIFO) stack providing push and pop(and peek) functionality.
-    // -----------------------------------------------------------------------------------------------------
-    // - Implemented Interfaces:
-    //    ICollection, IEnumerable, and ICloneable
-    //    
-
-    // ============================================================================================================
-    // Table 9-2. Key Interfaces Supported by Classes of System.Collections
-    // --------------------------------------------------------------------
-    // System.Collections Interface      Meaning in Life
-    // ============================================================================================================
-    // ICollection 
-    // ---------------
-    // Defines general characteristics (e.g., size, enumeration, and thread
-    // safety) for all nongeneric collection types
-    // ============================================================================================================
-    // ICloneable 
-    // ---------------
-    // Allows the implementing object to return a copy of itself to the caller
-    // ============================================================================================================
-    // IDictionary
-    // ---------------
-    // Allows a nongeneric collection object to represent its contents using key-value pairs
-    // ============================================================================================================
-    // IEnumerable 
-    // ---------------
-    // Returns an object implementing the IEnumerator interface (see next table entry)
-    // ============================================================================================================
-    // IEnumerator 
-    // ---------------
-    // Enables foreach-style iteration of collection items
-    // ============================================================================================================
-    // IList 
-    // ---------------
-    // Provides behavior to add, remove, and index items in a sequential list of objects
-    // ============================================================================================================
-
-    // Working with ArrayList
-    // ----------------------
-
-    //public class ArrayList : IList, ICollection, IEnumerable, ICloneable
-    //{
-    //    public ArrayList();
-    //    public ArrayList(int capacity);
-    //    public ArrayList(ICollection c);
-
-    //    public virtual object this[int index] { get; set; }
-
-    //    public virtual bool IsSynchronized { get; }
-    //    public virtual bool IsReadOnly { get; }
-    //    public virtual bool IsFixedSize { get; }
-    //    public virtual int Count { get; }
-    //    public virtual int Capacity { get; set; }
-    //    public virtual object SyncRoot { get; }
-
-    //    public static ArrayList Adapter(IList list);
-    //    public static IList FixedSize(IList list);
-    //    public static ArrayList FixedSize(ArrayList list);
-    //    public static IList ReadOnly(IList list);
-    //    public static ArrayList ReadOnly(ArrayList list);
-    //    public static ArrayList Repeat(object value, int count);
-    //    public static ArrayList Synchronized(ArrayList list);
-    //    public static IList Synchronized(IList list);
-    //    public virtual int Add(object value);
-    //    public virtual void AddRange(ICollection c);
-    //    public virtual int BinarySearch(int index, int count, object value, IComparer comparer);
-    //    public virtual int BinarySearch(object value);
-    //    public virtual int BinarySearch(object value, IComparer comparer);
-    //    public virtual void Clear();
-    //    public virtual object Clone();
-    //    public virtual bool Contains(object item);
-    //    public virtual void CopyTo(int index, Array array, int arrayIndex, int count);
-    //    public virtual void CopyTo(Array array);
-    //    public virtual void CopyTo(Array array, int arrayIndex);
-    //    public virtual IEnumerator GetEnumerator();
-    //    public virtual IEnumerator GetEnumerator(int index, int count);
-    //    public virtual ArrayList GetRange(int index, int count);
-    //    public virtual int IndexOf(object value, int startIndex, int count);
-    //    public virtual int IndexOf(object value, int startIndex);
-    //    public virtual int IndexOf(object value);
-    //    public virtual void Insert(int index, object value);
-    //    public virtual void InsertRange(int index, ICollection c);
-    //    public virtual int LastIndexOf(object value, int startIndex);
-    //    public virtual int LastIndexOf(object value, int startIndex, int count);
-    //    public virtual int LastIndexOf(object value);
-    //    public virtual void Remove(object obj);
-    //    public virtual void RemoveAt(int index);
-    //    public virtual void RemoveRange(int index, int count);
-    //    public virtual void Reverse(int index, int count);
-    //    public virtual void Reverse();
-    //    public virtual void SetRange(int index, ICollection c);
-    //    public virtual void Sort(IComparer comparer);
-    //    public virtual void Sort(int index, int count, IComparer comparer);
-    //    public virtual void Sort();
-    //    public virtual object[] ToArray();
-    //    public virtual Array ToArray(Type type);
-    //    public virtual void TrimToSize();
-    //}
-
-
-
-
-
-    // A Survey of System.Collections.Specialized Namespace
-    // =====================================================
-    // System.Collections is not the only.NET namespace that contains nongeneric collection classes.The
-    // System.Collections.Specialized namespace defines a number of(pardon the redundancy) specialized
-    // collection types.Table 9-3 documents some of the more useful types in this particular collection-centric
-    // namespace, all of which are nongeneric.
-    // Beyond these concrete class types, this namespace also contains many additional interfaces and
-    // abstract base classes that you can use as a starting point for creating custom collection classes.While these
-    // “specialized” types might be just what your projects require in some situations, I won’t comment on their
-    // usage here. Again, in many cases, you will likely find that the System.Collections.Generic namespace
-    // provides classes with similar functionality and additional benefits.
-    // ■ Note there are two additional collection-centric namespaces (System.Collections.ObjectModel and
-    // System.Collections.Concurrent) in the.net base class libraries. You will examine the former namespace
-    // later in this chapter, after you are comfortable with the topic of generics.System.Collections.Concurrent
-    // provides collection classes well-suited to a multithreaded environment (see Chapter 19 for information on
-    // multithreading).
-
-    // Table 9-3. Useful Classes of System.Collections.Specialized
-    // ==============================================================================================
-    // System.Collections.Specialized Type Meaning in Life
-    // ==============================================================================================
-    // HybridDictionary 
-    // ------------------------------------------------------------
-    // This class implements IDictionary by using a
-    // ListDictionary while the collection is small and then
-    // switching to a Hashtable when the collection gets large.
-    // ==============================================================================================
-    // ListDictionary
-    // ------------------------------------------------------------
-    // This class is useful when you need to manage a small
-    // number of items(ten or so) that can change over time.This
-    // class makes use of a singly linked list to maintain its data.
-    // ==============================================================================================
-    // StringCollection 
-    // ------------------------------------------------------------
-    // This class provides an optimal way to manage large
-    // collections of string data.
-    // ==============================================================================================
-    // BitVector32 
-    // ------------------------------------------------------------
-    // This class provides a simple structure that stores Boolean
-    // values and small integers in 32 bits of memory.
-
-    // ■ Note: there are two additional collection-centric namespaces(System.Collections.ObjectModel and
-    // System.Collections.Concurrent) in the.net base class libraries. You will examine the former namespace
-    // later in this chapter, after you are comfortable with the topic of generics.System.Collections.Concurrent
-    // provides collection classes well-suited to a multithreaded environment. 
-    #endregion
-
-    #region Working with the ArrayList Class
-
-    class WorkingWithArrayList
-    {
-        public static void Test()
-        {
-            Console.WriteLine();
-            Console.WriteLine("Working With ArrayList ");
-            Console.WriteLine("".PadLeft(40, '='));
-
-            ArrayList arList = new ArrayList();
-
-            arList.AddRange(new[] { "Moamen", "Mohammed", "Gamal" });
-
-            arList.Add("Soroor");
-
-            Console.Write("List Items {");
-            foreach (var item in arList)
-            {
-                Console.Write($" {item} ");
-            }
-            Console.WriteLine("}\n");
-
-
-        }
-
-    }
-
-    #endregion
-
-    #region Working with the Stack Class
-
-    class WorkingWithStack
-    {
-        public static void Test()
-        {
-            Console.WriteLine();
-            Console.WriteLine("Working With Stack ");
-            Console.WriteLine("".PadLeft(40, '='));
-
-            Stack arList = new Stack();
-
-
-
-        }
-
-    }
-
-    #endregion
-
-    #region Working with the Queue Class
-
-    class WorkingWithQueue
-    {
-        public static void Test()
-        {
-            Console.WriteLine();
-            Console.WriteLine("Working With Queue ");
-            Console.WriteLine("".PadLeft(40, '='));
-
-            Queue arList = new Queue();
-
-
-
-        }
-
-    }
-
-    #endregion
-
-    #region Working with the SortedList Class
-
-    class WorkingWithSortedList
-    {
-        public static void Test()
-        {
-            Console.WriteLine();
-            Console.WriteLine("Working With SortedList ");
-            Console.WriteLine("".PadLeft(40, '='));
-
-            SortedList arList = new SortedList();
-
-
-
-        }
-
-    }
-
-    #endregion
-
-    #region Working with the Hashtable Class
-
-    class WorkingWithHashtable
-    {
-        public static void Test()
-        {
-            Console.WriteLine();
-            Console.WriteLine("Working With Hashtable ");
-            Console.WriteLine("".PadLeft(40, '='));
-
-            Hashtable arList = new Hashtable();
-
-
-
-        }
-
-    }
-
-    #endregion
-
-    #region Working with the BitArray Class
-
-    class WorkingWithBitArray
-    {
-        public static void Test()
-        {
-            Console.WriteLine();
-            Console.WriteLine("Working With BitArray ");
-            Console.WriteLine("".PadLeft(40, '='));
-
-            BitArray arList = new BitArray(64);
-
-
-
-        }
-
-    }
-
-    #endregion
-
-    // ================================================
-    // The System.Collections.Specialized Namespace
-    // ================================================
-    // Requires System.Collections.Specialized;
-    #region Working with the HybridDictionary Class
-    class WorkingWithHybridDictionary
-    {
-        public static void Test()
-        {
-            Console.WriteLine();
-            Console.WriteLine("Working With HybridDictionary ");
-            Console.WriteLine("".PadLeft(40, '='));
-
-            HybridDictionary arList = new HybridDictionary();
-
-
-
-        }
-
-    }
-
-    #endregion
-
-    // Requires System.Collections.Specialized;
-    #region Working with the ListDictionary Class
-    // Requires System.Collections.Specialized
-    class WorkingWithListDictionary
-    {
-        public static void Test()
-        {
-            Console.WriteLine();
-            Console.WriteLine("Working With ListDictionary ");
-            Console.WriteLine("".PadLeft(40, '='));
-
-            ListDictionary arList = new ListDictionary();
-
-
-
-        }
-
-    }
-
-    #endregion
-
-    // Requires System.Collections.Specialized;
-    #region Working with the StringCollection Class
-    // Requires System.Collections.Specialized
-    class WorkingWithStringCollection
-    {
-        public static void Test()
-        {
-            Console.WriteLine();
-            Console.WriteLine("Working With StringCollection ");
-            Console.WriteLine("".PadLeft(40, '='));
-
-            StringCollection arList = new StringCollection();
-
-
-
-        }
-
-    }
-
-    #endregion
-
-    // Requires System.Collections.Specialized;
-    #region Working with the BitVector32 Class
-    // Requires System.Collections.Specialized
-    class WorkingWithBitVector32
-    {
-        public static void Test()
-        {
-            Console.WriteLine();
-            Console.WriteLine("Working With BitVector32 ");
-            Console.WriteLine("".PadLeft(40, '='));
-
-            BitVector32 arList = new BitVector32();
-
-
-
-        }
-
-    }
-
-    #endregion
-
-    
     #region Boxing and Unboxing
     // Boxing and Unboxing
     // =========================================================================================================
@@ -2684,7 +2256,7 @@ namespace ProCSharpCode.CSharpCollections
         // Summary:
         //     Represents a collection of key/value pairs that are sorted on the key.
         // it's exaclty like Dictionary<T> except that it is sorted and when you enumerate it 
-        // it will return it's data in the same order as it sorted
+        // it will return data sorted depends on the key
 
         public static void Test()
         {
@@ -2706,6 +2278,46 @@ namespace ProCSharpCode.CSharpCollections
             };
 
             PrintDictionary(sortedDictionary2, "Print SortedDictionary Items:");
+
+        }
+
+    }
+
+    #endregion
+
+    #region Working with the SortedList<T> Class
+
+    class WorkingWithGenericSortedList
+    {
+        // Summary: 
+        //     Represents a collection of key/value pairs that are sorted by key based
+        //     on the associated System.Collections.Generic.IComparer`1 implementation.
+        // it is functionaly the same as the SortedDictionary<T,K> except that
+        // it is internaly uses a List
+        // it is more memory efficient, but less in Modification that the SortedDictionary
+        // it's Modification O is O(n) but SortedDictionary<T,K> is O(n log n)
+
+
+        public static void Test()
+        {
+            Console.WriteLine();
+            Console.WriteLine("Working With Generic Sorted Dictionary");
+            Console.WriteLine("".PadLeft(40, '='));
+
+            // Create Object of SortedList With Object Init Syntax
+            // Using Default TKey Comparer
+            SortedList<string, Person> sortedDictionary2 = new SortedList<string, Person>()
+            {
+                // use of Object init Syntax as Dictionary<TKey, TValue> has Add() Function
+                { "Moamen" , new Person(104, "Moamen", 25)},
+                { "Mohammed" , new Person(103,"Mohammed",27)},
+                { "Kamal" , new Person(107,"Kamal",24)},
+                { "Mostafa" , new Person(101,"Mostafa",23)},
+                { "Ali" , new Person(102,"Ali",30)},
+                { "Waleed" , new Person(106,"Waleed",15)},
+            };
+
+            PrintDictionary(sortedDictionary2, "Print SortedList Items:");
 
         }
 
@@ -2805,10 +2417,66 @@ namespace ProCSharpCode.CSharpCollections
 
     #endregion
 
+
+    #region Working With ReadOnly Collections
+
+    // ================================================
+    // Working With ReadOnly Collections
+    // ================================================
+    // read only collections are just a wrapper of it's underlaying collection
+    // it consider as an interface/cover for the list
+    // - you can't add or remove items from the readonly collection
+    //   but you can add and remove from reference to it's underlaying collection
+    //   so it is not immutable but just a read only interface pattern
+
+    #endregion
+
+
+    // namespace of it is System.Collections.ObjectModel;
+    #region Working with the ReadOnlyCollection<T> Class
+    // namespace of it is System.Collections.ObjectModel;
+    // Provides the base class for a generic read-only collection.
+    // 
+    public class WorkingWithGenericReadOnlyCollection
+    {
+
+
+        public static void Test()
+        {
+            Console.WriteLine();
+            Console.WriteLine("Working With Generic ReadOnlyCollection");
+            Console.WriteLine("".PadLeft(40, '='));
+
+            // Create Object of ObservableCollection
+            List<string> list = new List<string>() { "ali", "ahmed", "mohammed" };
+
+            ReadOnlyCollection<string> readonlyList = new ReadOnlyCollection<string>(list);
+
+            // now you can't add or remove using readonlyList
+            //readonlyList.Add(); // add not exists
+
+            // but you can add or remove to the underlaying collection
+            // and that modification will reflect on the ReadOnlyCollection
+            list.Add("Mahmmoud");
+
+            Console.WriteLine(string.Join(", ", readonlyList));
+
+            // Expected output: 
+            // Working With Generic ReadOnlyCollection
+            // ========================================
+            // ali, ahmed, mohammed, Mahmmoud
+
+        }
+
+    }
+
+    #endregion
+
+
     // namespace of it is System.Collections.ObjectModel;
     #region Working with the ReadOnlyObservableCollection<T> Class
     // namespace of it is System.Collections.ObjectModel;
-    class WorkingWithGenericReadOnlyObservableCollection
+    public class WorkingWithGenericReadOnlyObservableCollection
     {
 
 
@@ -2828,6 +2496,90 @@ namespace ProCSharpCode.CSharpCollections
     }
 
     #endregion
+
+
+    #region Working With Immutable Collections
+    // ================================================
+    // Working With Immutable Collections
+    // ================================================
+    // Immutable collections: it is doesn't allow modification to be done in the same object
+    //  all the modification methods returns a new list with the update in it
+    //  without change the current collection
+
+    // IImmutableList<> Interface has the following methods: 
+
+    //IImmutableList<T> Add(T value);
+    //IImmutableList<T> AddRange(IEnumerable<T> items);
+    //IImmutableList<T> Clear();
+    //int IndexOf(T item, int index, int count, IEqualityComparer<T>? equalityComparer);
+    //IImmutableList<T> Insert(int index, T element);
+    //IImmutableList<T> InsertRange(int index, IEnumerable<T> items);
+    //int LastIndexOf(T item, int index, int count, IEqualityComparer<T>? equalityComparer);
+    //IImmutableList<T> Remove(T value, IEqualityComparer<T>? equalityComparer);
+    //IImmutableList<T> RemoveAll(Predicate<T> match);
+    //IImmutableList<T> RemoveAt(int index);
+    //IImmutableList<T> RemoveRange(IEnumerable<T> items, IEqualityComparer<T>? equalityComparer);
+    //IImmutableList<T> RemoveRange(int index, int count);
+    //IImmutableList<T> Replace(T oldValue, T newValue, IEqualityComparer<T>? equalityComparer);
+    //IImmutableList<T> SetItem(int index, T value);
+
+
+
+    //public ImmutableList<T>.Builder ToBuilder();
+
+    public class WorkingWithImmutableCollections
+    {
+
+        public static void Test()
+        {
+            Console.WriteLine();
+            Console.WriteLine("Working With ImmutableCollections");
+            Console.WriteLine("".PadLeft(40, '='));
+
+
+            // Create Object of ObservableCollection
+            List<string> list = new List<string>() { "ali", "ahmed", "mohammed" };
+
+            ImmutableList<string> imList = list.ToImmutableList();
+            var newImList = imList.Add("ahmed");
+
+            var newImList2 = newImList.Remove("ali");
+
+
+            Console.WriteLine($"imList----: {string.Join(", ", imList)}");
+            Console.WriteLine($"newImList-: {string.Join(", ", newImList)}");
+            Console.WriteLine($"newImList2: {string.Join(", ", newImList2)}");
+
+            // Expected output: 
+            // Working With ImmutableCollections
+            // ========================================
+            // imList----: ali, ahmed, mohammed
+            // newImList-: ali, ahmed, mohammed, ahmed
+            // newImList2: ahmed, mohammed, ahmed
+
+
+
+            // immutable types that are the immutable version of mutable collections
+
+            ImmutableArray<string> imArray = list.ToImmutableArray();
+            ImmutableHashSet<string> imHashset = list.ToImmutableHashSet();
+            ImmutableDictionary<string,string> imDictionary = list.ToImmutableDictionary(k=> k);
+            ImmutableSortedDictionary<string,string> imSortedDictionary = list.ToImmutableSortedDictionary(k=> k,v=> v);
+            ImmutableSortedSet<string> imSortedSet = list.ToImmutableSortedSet(StringComparer.OrdinalIgnoreCase);
+
+
+
+        }
+
+    }
+
+
+
+
+    #endregion
+
+
+
 
     class CollectionsTraining
     {
